+++ .gitignore
@@ -3,7 +3,6 @@
 # Python files
 __pycache__/
 # Example files
-example_files/
+example_files/asperaweb_id_dsa.openssh
 data/
 tmp/
-idr_stream.log
