+++ README.md
@@ -1,2 +1,25 @@
-# IDR_stream
+# idrstream
+
 Software for feature extraction from IDR image data
+
+250 min, 87 batches * 10 images/batch = 870 images, 1.3 GB
+
+## About
+
+Idrstream is a package for deriving features from medical image data hosted on [IDR](https://idr.openmicroscopy.org/).
+The stream processes image data in the following pipeline:
+
+![Stream Design](images/Stream_Design.png "Stream Design")
+
+
+The stream functions as follows:
+
+1) Create the base Deep Profiler 
+
+## Setup
+
+
+
+## Usage
+
+
