+++ example_files/DP_files/mitocheck_profiling_config.json
@@ -1,0 +1,84 @@
+{
+    "dataset": {
+        "metadata": {
+            "label_field": "Gene",
+            "control_value": "EMPTY"
+        },
+        "images": {
+            "channels": [
+                "DNA"
+            ],
+            "file_format": "tif",
+            "bits": 8,
+            "width": 1344,
+            "height": 1024
+        },
+        "locations": {
+            "mode": "single_cells",
+            "box_size": 128,
+            "area_coverage": 0.75,
+            "mask_objects": false
+        }
+    },
+    "prepare": {
+        "illumination_correction": {
+            "down_scale_factor": 4,
+            "median_filter_size": 24
+        },
+        "compression": {
+            "implement": false,
+            "scaling_factor": 1.0
+        }
+    },
+    "train": {
+        "partition": {
+            "targets": [
+                "Gene"
+            ],
+            "split_field": "Gene_Replicate",
+            "training_values": [
+                0
+            ],
+            "validation_values": [
+                1
+            ]
+        },
+        "model": {
+            "name": "efficientnet",
+            "crop_generator": "repeat_channel_crop_generator",
+            "metrics": [
+                "accuracy",
+                "top_k"
+            ],
+            "epochs": 100,
+            "initialization": "ImageNet",
+            "params": {
+            	"label_smoothing": 0.0,
+                "learning_rate": 0.005,
+                "batch_size": 64,
+                "conv_blocks": 0,
+                "feature_dim": 256,
+                "pooling": "avg"
+            },
+            "lr_schedule": "cosine"
+        },
+        "sampling": {
+            "factor": 1,
+            "workers": 4,
+            "cache_size": 15000
+        },
+        "validation": {
+            "frequency": 2,
+            "top_k": 5,
+            "batch_size": 32,
+            "frame": "val",
+            "sample_first_crops": true
+        }
+    },
+    "profile": {
+        "use_pretrained_input_size": 224,
+        "feature_layer": "avg_pool",
+        "checkpoint": "efficientnet-b0_weights_tf_dim_ordering_tf_kernels_autoaugment.h5",
+        "batch_size": 1024
+    }
+}
